+++ addons/pt/saft/invoices.go
@@ -10,6 +10,7 @@ import (
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cal"
 	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/rules"
 	"github.com/invopop/gobl/rules/is"
@@ -96,6 +97,7 @@ func normalizeInvoiceValueDate(inv *bill.Invoice) {
 
 func billInvoiceRules() *rules.Set {
 	return rules.For(new(bill.Invoice),
+		rules.Assert("15", "invoice must be in EUR or provide exchange rate for conversion", currency.CanConvertTo(currency.EUR)),
 		// Tax ext must have either work type or invoice type
 		rules.Assert("01",
 			fmt.Sprintf("either '%s' or '%s' must be set", ExtKeyWorkType, ExtKeyInvoiceType),